requirements.txt
numpy>=1.21.0
tensorflow>=2.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
cryptography>=3.4.8
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0
jupyter>=1.0.0
plotly>=5.0.0
pyyaml>=6.0
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
alembic>=1.7.0
pytest>=6.0.0
pytest-cov>=3.0.0
pytest-asyncio>=0.19.0
pytest-benchmark>=3.4.0
black>=21.0.0
flake8>=4.0.0
mypy>=0.910
pre-commit>=2.15.0
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0